requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-appconf==1.0.4
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-redis==4.11.0
django-select2==7.2.3
django-timezone-field==4.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
honcho==1.0.1
joblib==0.14.1
kombu==4.6.8
numpy==1.18.4
pandas==1.0.3
psycopg2-binary==2.8.5
pycparser==2.20
python-crontab==2.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
static3==0.7.0
vine==1.3.0
whitenoise==5.0.1
xgboost==1.0.2
